requirements.txt
h5py==3.7.0
open3d==0.15.2
pytorch-lightning==1.6.4
lightning-bolts==0.5.0
PyYAML==6.0
timm==0.6.7
torch-cka==0.21
tqdm==4.64.0
wandb==0.12.19
ninja==1.10.2.3
torch-geometric==2.0.4
torch-scatter==2.0.9
torch-sparse==0.6.13
torchmetrics==0.9.1
